Move autoload to main.ts
