+++ src/main.ts
@@ -38,6 +38,8 @@ import {LogicalGraph} from './LogicalGraph';
 import {Modals} from './Modals';
 import {Palette} from './Palette';
 import {Utils} from './Utils';
+import {Repository} from './Repository';
+import {RepositoryFile} from './RepositoryFile';
 
 let eagle : Eagle;
 
@@ -118,5 +120,89 @@ $(function(){
     const service: Eagle.RepositoryService = Eagle.RepositoryService[auto_load_service as keyof typeof Eagle.RepositoryService];
 
     // auto load the file
-    eagle.autoLoad(service, auto_load_repository, auto_load_branch, auto_load_path, auto_load_filename);
+    autoLoad(eagle, service, auto_load_repository, auto_load_branch, auto_load_path, auto_load_filename);
+
+    //hides the dropdown navbar elements when stopping hovering over the element
+    $(".dropdown-menu").mouseleave(function(){
+        $(".dropdown-toggle").removeClass("show")
+        $(".dropdown-menu").removeClass("show")
+      })
+  
+      $('.modal').on('hidden.bs.modal', function () {
+          $('.modal-dialog').css({"left":"0px", "top":"0px"})
+          $("#editFieldModal textarea").attr('style','')
+          $("#errorsModalAccordion").parent().parent().attr('style','')
+  
+          //reset parameter table selecction
+          Eagle.resetParamsTableSelection()
+      });
+  
+      $('.modal').on('shown.bs.modal',function(){
+          // modal draggables
+          //the any type is required so we dont have an error when building. at runtime on eagle this actually functions without it.
+          (<any>$('.modal-dialog')).draggable({
+              handle: ".modal-header"
+          });
+      })
+  
+      //increased click bubble for edit modal flag booleans
+      $(".componentCheckbox").on("click",function(){
+          $(event.target).find("input").click()
+      })
+  
+      $('#editFieldModalValueInputCheckbox').on("change",function(){
+          $(event.target).parent().find("span").text($(event.target).prop('checked'))
+      })
+  
+      //removes focus from input and textareas when using the canvas
+      $("#logicalGraphParent").on("mousedown", function(){
+          $("input").blur();
+          $("textarea").blur();
+      });
+  
+      $(".tableParameter").on("click", function(){
+          console.log(this)
+      })
+  
+      //expand palettes when using searchbar and return to prior collapsed state on completion.
+      $("#paletteList .componentSearchBar").on("keyup",function(){
+          if ($("#paletteList .componentSearchBar").val() !== ""){
+              $("#paletteList .accordion-button.collapsed").addClass("wasCollapsed")
+              $("#paletteList .accordion-button.collapsed").click()
+          }else{
+              $("#paletteList .accordion-button.wasCollapsed").click()
+              $("#paletteList .accordion-button.wasCollapsed").removeClass("wasCollapsed")
+          }
+      })
+  
+      $(document).on('click', '.hierarchyEdgeExtra', function(){
+          const selectEdge = (<any>window).eagle.logicalGraph().findEdgeById(($(event.target).attr("id")))
+  
+          if(!selectEdge){
+              console.log("no edge found")
+              return
+          }
+          if(!(<PointerEvent>event).shiftKey){
+              (<any>window).eagle.setSelection(Eagle.RightWindowMode.Inspector, selectEdge, Eagle.FileType.Graph);
+          }else{
+              (<any>window).eagle.editSelection(Eagle.RightWindowMode.Inspector, selectEdge, Eagle.FileType.Graph);
+          }
+  
+      })
+      $(".hierarchy").on("click", function(){
+          (<any>window).eagle.selectedObjects([]);
+      })
 });
+
+function autoLoad(eagle: Eagle, service: Eagle.RepositoryService, repository: string, branch: string, path: string, filename: string) {
+    console.log("autoLoad()", service, repository, branch, path, filename);
+
+    // skip empty string urls
+    if (service === Eagle.RepositoryService.Unknown || repository === "" || branch === "" || filename === ""){
+        console.log("No auto load");
+        return;
+    }
+
+    // load
+    eagle.selectFile(new RepositoryFile(new Repository(service, repository, branch, false), path, filename));
+}
